move valid solvent check to utilities
